Fix reference problem
